Check revisions before rebuild

+++ generate.sh
@@ -19,6 +19,14 @@ else
     git clone --depth 1 https://github.com/rurema/doctree repos/doctree
 fi
 
+git -C repos/bitclust rev-parse HEAD > db/bitclust.rev
+git -C repos/doctree rev-parse HEAD > db/doctree.rev
+
+if [ -z "$(git status -s db/*.rev)" ]; then
+    echo do nothing because bitclust and doctree are same revisions
+    exit
+fi
+
 time docker compose build rurema
 
 time docker compose run --rm rurema ls -al
